Bugfixes and feature additions, see below
- Feature: amounts for items
- Feature: categorize items and define category walk order for a store
- Improve: page titles are now actually unique per page
- Improve: color coding now makes very similar strings a similar color
- Improve: popularItems would often update too early since moving to a
  new server, giving you the state from *before* the item was
  added/removed to/from the list. A small delay is introduced.
- Fix that items with an apostrophe, ampersand, etc. would not be
  handled correctly
- Fix a potential XSS (str.replace("'", ...) needed a /g flag)

+++ processRequests.php
@@ -15,6 +15,21 @@ if ($_SESSION['loggedin'] === 'yes') {
 			exit;
 		}
 
+		if (isset($_GET['catorder'])) {
+			$store = intval($_POST['store']);
+
+			foreach ($_POST as $key=>$val) {
+				if (substr($key, 0, 5) == 'catid') {
+					$catid = intval(substr($key, 5));
+					$prio = floatval($val);
+					$db->query("INSERT INTO categories_stores (priority, storeid, categoryid, uid) VALUES($prio, $store, $catid, $_SESSION[uid])
+						ON DUPLICATE KEY UPDATE priority = $prio") or die('Database error 737594');
+				}
+			}
+			header('Location: ?map&store=' . $store);
+			exit;
+		}
+
 		if (isset($_GET['dontPurchase'])) {
 			$itemstoreid = intval($_GET['dontPurchase']);
 			$db->query("DELETE FROM item_stores WHERE uid = $_SESSION[uid] AND id = $itemstoreid") or die('Database error 6446');
